content/canvas/test/webgl-mochitest/test_privileged_exts.html

<!DOCTYPE HTML>
<html>
<head>
<title>WebGL test: Check for privileged ext access.</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<script src="webgl-util.js"></script>
<script src="driver-info.js"></script>
</head>
<body>
<canvas id="c"></canvas>
<script>

function TestExt(gl, name) {
  var ext = gl.getExtension(name);
  ok(!ext, 'Should not have access to \'' + name + '\'.');
}

(function() {
  var gl = WebGLUtil.getWebGL('c');
  if (!gl) {
    todo(gl, 'Get GL working here first.');
    return;
  }

  // Privileged extensions:
  TestExt(gl, 'WEBGL_debug_renderer_info');
})();

</script>
</body>
</html>
